+++ lib/praxis-blueprints.rb
@@ -12,3 +12,6 @@ require 'praxis-blueprints/config_hash'
 require 'praxis-blueprints/blueprint'
 require 'praxis-blueprints/view'
 require 'praxis-blueprints/collection_view'
+
+require 'praxis-blueprints/renderer'
+require 'praxis-blueprints/field_expander'