+++ lib/praxis-blueprints/view.rb
@@ -6,18 +6,13 @@ module Praxis
     attr_reader :name
     attr_reader :options
 
-    attr_reader :include_nil
-
     def initialize(name, schema, **options, &block)
       @name = name
       @schema = schema
       @contents = ::Hash.new
       @block = block
 
-      @include_nil = options.fetch(:include_nil, false)
-
       @options = options
-
     end
 
     def contents
@@ -29,52 +24,64 @@ module Praxis
       @contents
     end
 
-    def dump(object, context: Attributor::DEFAULT_ROOT_CONTEXT,**opts)
-      fields = opts[:fields]
-      # Restrict which attributes to output if we receive a fields parameter
-      # Note: should we complain if any of the names in "fields" do not match an existing attribute name?
-      attributes_to_render = self.contents.keys
-      attributes_to_render &= fields.keys if fields
-
-      attributes_to_render.each_with_object({}) do |name, hash|
-        dumpable, dumpable_opts = self.contents[name]
-        unless object.respond_to?(name)
-          warn "#{object} does not respond to #{name} during rendering???"
-          next
-        end
-
-        begin
-          value = object.send(name)
-        rescue => e
-          raise Attributor::DumpError, context: context, name: name, type: object.class, original_exception: e
-        end
-
-        if value.nil?
-          next unless @include_nil
-        end
-
-        # FIXME: this is such an ugly way to do this. Need attributor#67.
-        if dumpable.kind_of?(View) || dumpable.kind_of?(CollectionView)
-          new_context = context + [name]
-
-          sub_opts = add_subfield_options( fields, name, dumpable_opts )
-          hash[name] = dumpable.dump(value, context: new_context ,**sub_opts)
-        else
-
-          type = dumpable.type
-          if type.respond_to?(:attributes) || type.respond_to?(:member_attribute)
-            new_context = context + [name]
-            sub_opts = add_subfield_options( fields, name, dumpable_opts )
-            hash[name] = dumpable.dump(value, context: new_context ,**sub_opts)
-          else
-            hash[name] = value
-          end
-        end
+    def expanded_fields
+      @expanded_fields ||= begin
+        self.contents # force evaluation of the contents
+        FieldExpander.expand(self)
       end
     end
-    alias_method :to_hash, :dump
 
-    def attribute(name, opts={}, &block)
+    def render(object, context: Attributor::DEFAULT_ROOT_CONTEXT, renderer: Renderer.new)
+      renderer.render(object, self.expanded_fields, context: context)
+    end
+
+    alias_method :to_hash, :render # Why did we need this again?
+
+    # def dump(object, context: Attributor::DEFAULT_ROOT_CONTEXT,**opts)
+    #   fields = opts[:fields]
+    #   # Restrict which attributes to output if we receive a fields parameter
+    #   # Note: should we complain if any of the names in "fields" do not match an existing attribute name?
+    #   attributes_to_render = self.contents.keys
+    #   attributes_to_render &= fields.keys if fields
+    #
+    #   attributes_to_render.each_with_object({}) do |name, hash|
+    #     dumpable, dumpable_opts = self.contents[name]
+    #     unless object.respond_to?(name)
+    #       warn "#{object} does not respond to #{name} during rendering???"
+    #       next
+    #     end
+    #
+    #     begin
+    #       value = object.send(name)
+    #     rescue => e
+    #       raise Attributor::DumpError, context: context, name: name, type: object.class, original_exception: e
+    #     end
+    #
+    #     if value.nil?
+    #       next unless @include_nil
+    #     end
+    #
+    #     # FIXME: this is such an ugly way to do this. Need attributor#67.
+    #     if dumpable.kind_of?(View) || dumpable.kind_of?(CollectionView)
+    #       new_context = context + [name]
+    #
+    #       sub_opts = add_subfield_options(fields, name, dumpable_opts)
+    #       hash[name] = dumpable.dump(value, context: new_context, **sub_opts)
+    #     else
+    #       type = dumpable.type
+    #       if type.respond_to?(:attributes) || type.respond_to?(:member_attribute)
+    #         new_context = context + [name]
+    #         sub_opts = add_subfield_options(fields, name, dumpable_opts)
+    #         hash[name] = dumpable.dump(value, context: new_context ,**sub_opts)
+    #       else
+    #         hash[name] = value
+    #       end
+    #     end
+    #   end
+    # end
+
+
+    def attribute(name, **opts, &block)
       raise AttributorException, "Attribute names must be symbols, got: #{name.inspect}" unless name.kind_of? ::Symbol
 
       attribute = self.schema.attributes.fetch(name) do
@@ -82,42 +89,60 @@ module Praxis
       end
 
       if block_given?
-        view = View.new(name, attribute, &block)
-        @contents[name] = view
+        type = attribute.type
+        @contents[name] = if type < Attributor::Collection
+          CollectionView.new(name, type.member_attribute.type, &block)
+        else
+          View.new(name, attribute, &block)
+        end
       else
-        raise "Invalid options (#{opts.inspect}) for #{name} while defining view #{@name}" unless opts.is_a?(Hash)
-        @contents[name] = [attribute, opts]
+        type = attribute.type
+        if type < Attributor::Collection
+          is_collection = true
+          type = type.member_attribute.type
+        end
+
+
+        if type < Praxis::Blueprint
+          view_name = opts[:view] || :default
+          view = type.views.fetch(view_name) do
+            raise "view with name '#{view_name.inspect}' is not defined in #{type}"
+          end
+          if is_collection
+            @contents[name] = Praxis::CollectionView.new(view_name, type, view)
+          else
+            @contents[name] = view
+          end
+        else
+          @contents[name] = attribute #, opts]
+        end
       end
 
     end
 
-    def example(context=nil)
+    def example(context=Attributor::DEFAULT_ROOT_CONTEXT)
       object = self.schema.example(context)
       opts = {}
       opts[:context] = context if context
-      self.dump(object, opts)
+      self.render(object, opts)
     end
 
     def describe
       # TODO: for now we are just return the first level keys
       view_attributes = {}
 
-      self.contents.each do |k,(dumpable,dumpable_opts)|
+      self.contents.each do |k,dumpable|
         inner_desc = {}
-        inner_desc[:view] = dumpable_opts[:view] if dumpable_opts && dumpable_opts[:view]
+        #inner_desc[:view] = dumpable_opts[:view] if dumpable_opts && dumpable_opts[:view]
+        if dumpable.kind_of?(Praxis::View)
+          inner_desc[:view] = dumpable.name if dumpable.name
+        end
         view_attributes[k] = inner_desc
       end
 
       { attributes: view_attributes, type: :standard }
     end
 
-    private
-    def add_subfield_options(fields, name, existing_options)
-      sub_opts = if fields && fields[name]
-        {fields: fields[name] }
-      else
-        {}
-      end.merge(existing_options || {})
-    end
+
   end
 end